refactor: move server config defaults to pkg/server/config
