+++ cmd/run/run_test.go
@@ -31,6 +31,7 @@ import (
 	"github.com/openfga/openfga/cmd/util"
 	"github.com/openfga/openfga/internal/mocks"
 	"github.com/openfga/openfga/pkg/logger"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	serverErrors "github.com/openfga/openfga/pkg/server/errors"
 	"github.com/openfga/openfga/pkg/tuple"
 	"github.com/openfga/openfga/pkg/typesystem"
@@ -223,7 +224,7 @@ type authTest struct {
 	expectedStatusCode    int
 }
 
-func runServer(ctx context.Context, cfg *Config) error {
+func runServer(ctx context.Context, cfg *serverconfig.Config) error {
 	if err := VerifyConfig(cfg); err != nil {
 		return err
 	}
@@ -233,105 +234,10 @@ func runServer(ctx context.Context, cfg *Config) error {
 	return serverCtx.Run(ctx, cfg)
 }
 
-func TestVerifyConfig(t *testing.T) {
-	t.Run("UpstreamTimeout_cannot_be_less_than_ListObjectsDeadline", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.ListObjectsDeadline = 5 * time.Minute
-		cfg.HTTP.UpstreamTimeout = 2 * time.Second
-
-		err := VerifyConfig(cfg)
-		require.EqualError(t, err, "config 'http.upstreamTimeout' (2s) cannot be lower than 'listObjectsDeadline' config (5m0s)")
-	})
-
-	t.Run("failing_to_set_http_cert_path_will_not_allow_server_to_start", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.HTTP.TLS = &TLSConfig{
-			Enabled: true,
-			KeyPath: "some/path",
-		}
-
-		err := VerifyConfig(cfg)
-		require.EqualError(t, err, "'http.tls.cert' and 'http.tls.key' configs must be set")
-	})
-
-	t.Run("failing_to_set_grpc_cert_path_will_not_allow_server_to_start", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.GRPC.TLS = &TLSConfig{
-			Enabled: true,
-			KeyPath: "some/path",
-		}
-
-		err := VerifyConfig(cfg)
-		require.EqualError(t, err, "'grpc.tls.cert' and 'grpc.tls.key' configs must be set")
-	})
-
-	t.Run("failing_to_set_http_key_path_will_not_allow_server_to_start", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.HTTP.TLS = &TLSConfig{
-			Enabled:  true,
-			CertPath: "some/path",
-		}
-
-		err := VerifyConfig(cfg)
-		require.EqualError(t, err, "'http.tls.cert' and 'http.tls.key' configs must be set")
-	})
-
-	t.Run("failing_to_set_grpc_key_path_will_not_allow_server_to_start", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.GRPC.TLS = &TLSConfig{
-			Enabled:  true,
-			CertPath: "some/path",
-		}
-
-		err := VerifyConfig(cfg)
-		require.EqualError(t, err, "'grpc.tls.cert' and 'grpc.tls.key' configs must be set")
-	})
-
-	t.Run("non_log_format", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.Log.Format = "notaformat"
-
-		err := VerifyConfig(cfg)
-		require.Error(t, err)
-	})
-
-	t.Run("non_log_level", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.Log.Level = "notalevel"
-
-		err := VerifyConfig(cfg)
-		require.Error(t, err)
-	})
-
-	t.Run("empty_request_duration_datastore_query_count_buckets", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.RequestDurationDatastoreQueryCountBuckets = []string{}
-
-		err := VerifyConfig(cfg)
-		require.Error(t, err)
-	})
-
-	t.Run("non_int_request_duration_datastore_query_count_buckets", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.RequestDurationDatastoreQueryCountBuckets = []string{"12", "45a", "66"}
-
-		err := VerifyConfig(cfg)
-		require.Error(t, err)
-	})
-
-	t.Run("negative_request_duration_datastore_query_count_buckets", func(t *testing.T) {
-		cfg := DefaultConfig()
-		cfg.RequestDurationDatastoreQueryCountBuckets = []string{"12", "-45", "66"}
-
-		err := VerifyConfig(cfg)
-		require.Error(t, err)
-	})
-}
-
 func TestBuildServiceWithPresharedKeyAuthenticationFailsIfZeroKeys(t *testing.T) {
 	cfg := MustDefaultConfigWithRandomPorts()
 	cfg.Authn.Method = "preshared"
-	cfg.Authn.AuthnPresharedKeyConfig = &AuthnPresharedKeyConfig{}
+	cfg.Authn.AuthnPresharedKeyConfig = &serverconfig.AuthnPresharedKeyConfig{}
 
 	err := runServer(context.Background(), cfg)
 	require.EqualError(t, err, "failed to initialize authenticator: invalid auth configuration, please specify at least one key")
@@ -364,7 +270,7 @@ func TestBuildServiceWithNoAuth(t *testing.T) {
 func TestBuildServiceWithPresharedKeyAuthentication(t *testing.T) {
 	cfg := MustDefaultConfigWithRandomPorts()
 	cfg.Authn.Method = "preshared"
-	cfg.Authn.AuthnPresharedKeyConfig = &AuthnPresharedKeyConfig{
+	cfg.Authn.AuthnPresharedKeyConfig = &serverconfig.AuthnPresharedKeyConfig{
 		Keys: []string{"KEYONE", "KEYTWO"},
 	}
 
@@ -551,7 +457,7 @@ func tryGetStores(t *testing.T, test authTest, httpAddr string, retryClient *ret
 func TestHTTPServerWithCORS(t *testing.T) {
 	cfg := MustDefaultConfigWithRandomPorts()
 	cfg.Authn.Method = "preshared"
-	cfg.Authn.AuthnPresharedKeyConfig = &AuthnPresharedKeyConfig{
+	cfg.Authn.AuthnPresharedKeyConfig = &serverconfig.AuthnPresharedKeyConfig{
 		Keys: []string{"KEYONE", "KEYTWO"},
 	}
 	cfg.HTTP.CORSAllowedOrigins = []string{"http://openfga.dev", "http://localhost"}
@@ -650,7 +556,7 @@ func TestBuildServerWithOIDCAuthentication(t *testing.T) {
 
 	cfg := MustDefaultConfigWithRandomPorts()
 	cfg.Authn.Method = "oidc"
-	cfg.Authn.AuthnOIDCConfig = &AuthnOIDCConfig{
+	cfg.Authn.AuthnOIDCConfig = &serverconfig.AuthnOIDCConfig{
 		Audience: "openfga.dev",
 		Issuer:   localOIDCServerURL,
 	}
@@ -718,7 +624,7 @@ func TestHTTPServingTLS(t *testing.T) {
 		defer certsAndKeys.Clean()
 
 		cfg := MustDefaultConfigWithRandomPorts()
-		cfg.HTTP.TLS = &TLSConfig{
+		cfg.HTTP.TLS = &serverconfig.TLSConfig{
 			CertPath: certsAndKeys.serverCertFile,
 			KeyPath:  certsAndKeys.serverKeyFile,
 		}
@@ -740,7 +646,7 @@ func TestHTTPServingTLS(t *testing.T) {
 		defer certsAndKeys.Clean()
 
 		cfg := MustDefaultConfigWithRandomPorts()
-		cfg.HTTP.TLS = &TLSConfig{
+		cfg.HTTP.TLS = &serverconfig.TLSConfig{
 			Enabled:  true,
 			CertPath: certsAndKeys.serverCertFile,
 			KeyPath:  certsAndKeys.serverKeyFile,
@@ -778,7 +684,7 @@ func TestGRPCServingTLS(t *testing.T) {
 
 		cfg := MustDefaultConfigWithRandomPorts()
 		cfg.HTTP.Enabled = false
-		cfg.GRPC.TLS = &TLSConfig{
+		cfg.GRPC.TLS = &serverconfig.TLSConfig{
 			CertPath: certsAndKeys.serverCertFile,
 			KeyPath:  certsAndKeys.serverKeyFile,
 		}
@@ -801,7 +707,7 @@ func TestGRPCServingTLS(t *testing.T) {
 
 		cfg := MustDefaultConfigWithRandomPorts()
 		cfg.HTTP.Enabled = false
-		cfg.GRPC.TLS = &TLSConfig{
+		cfg.GRPC.TLS = &serverconfig.TLSConfig{
 			Enabled:  true,
 			CertPath: certsAndKeys.serverCertFile,
 			KeyPath:  certsAndKeys.serverKeyFile,
