+++ internal/graph/check.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"sync"
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 	"github.com/openfga/openfga/internal/validation"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/storage/storagewrappers"
 	"github.com/openfga/openfga/pkg/tuple"
@@ -20,12 +20,6 @@ import (
 )
 
 var tracer = otel.Tracer("internal/graph/check")
-
-const (
-	// same values as run.DefaultConfig() (TODO break the import cycle, remove these hardcoded values and import those constants here)
-	defaultResolveNodeBreadthLimit    = 25
-	defaultMaxConcurrentReadsForCheck = math.MaxUint32
-)
 
 var (
 	ErrCycleDetected = errors.New("a cycle has been detected")
@@ -158,8 +152,8 @@ func WithCachedResolver(opts ...CachedCheckResolverOpt) LocalCheckerOption {
 func NewLocalChecker(ds storage.RelationshipTupleReader, opts ...LocalCheckerOption) CheckResolver {
 	checker := &LocalChecker{
 		ds:                 ds,
-		concurrencyLimit:   defaultResolveNodeBreadthLimit,
-		maxConcurrentReads: defaultMaxConcurrentReadsForCheck,
+		concurrencyLimit:   serverconfig.DefaultResolveNodeBreadthLimit,
+		maxConcurrentReads: serverconfig.DefaultMaxConcurrentReadsForCheck,
 	}
 	checker.delegate = checker // by default, a LocalChecker delegates/dispatchs subproblems to itself (e.g. local dispatch) unless otherwise configured.
