+++ pkg/server/config/config_test.go
@@ -1,0 +1,103 @@
+package config
+
+import (
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestVerifyConfig(t *testing.T) {
+	t.Run("UpstreamTimeout_cannot_be_less_than_ListObjectsDeadline", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.ListObjectsDeadline = 5 * time.Minute
+		cfg.HTTP.UpstreamTimeout = 2 * time.Second
+
+		err := cfg.Verify()
+		require.EqualError(t, err, "config 'http.upstreamTimeout' (2s) cannot be lower than 'listObjectsDeadline' config (5m0s)")
+	})
+
+	t.Run("failing_to_set_http_cert_path_will_not_allow_server_to_start", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.HTTP.TLS = &TLSConfig{
+			Enabled: true,
+			KeyPath: "some/path",
+		}
+
+		err := cfg.Verify()
+		require.EqualError(t, err, "'http.tls.cert' and 'http.tls.key' configs must be set")
+	})
+
+	t.Run("failing_to_set_grpc_cert_path_will_not_allow_server_to_start", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.GRPC.TLS = &TLSConfig{
+			Enabled: true,
+			KeyPath: "some/path",
+		}
+
+		err := cfg.Verify()
+		require.EqualError(t, err, "'grpc.tls.cert' and 'grpc.tls.key' configs must be set")
+	})
+
+	t.Run("failing_to_set_http_key_path_will_not_allow_server_to_start", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.HTTP.TLS = &TLSConfig{
+			Enabled:  true,
+			CertPath: "some/path",
+		}
+
+		err := cfg.Verify()
+		require.EqualError(t, err, "'http.tls.cert' and 'http.tls.key' configs must be set")
+	})
+
+	t.Run("failing_to_set_grpc_key_path_will_not_allow_server_to_start", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.GRPC.TLS = &TLSConfig{
+			Enabled:  true,
+			CertPath: "some/path",
+		}
+
+		err := cfg.Verify()
+		require.EqualError(t, err, "'grpc.tls.cert' and 'grpc.tls.key' configs must be set")
+	})
+
+	t.Run("non_log_format", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.Log.Format = "notaformat"
+
+		err := cfg.Verify()
+		require.Error(t, err)
+	})
+
+	t.Run("non_log_level", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.Log.Level = "notalevel"
+
+		err := cfg.Verify()
+		require.Error(t, err)
+	})
+
+	t.Run("empty_request_duration_datastore_query_count_buckets", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.RequestDurationDatastoreQueryCountBuckets = []string{}
+
+		err := cfg.Verify()
+		require.Error(t, err)
+	})
+
+	t.Run("non_int_request_duration_datastore_query_count_buckets", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.RequestDurationDatastoreQueryCountBuckets = []string{"12", "45a", "66"}
+
+		err := cfg.Verify()
+		require.Error(t, err)
+	})
+
+	t.Run("negative_request_duration_datastore_query_count_buckets", func(t *testing.T) {
+		cfg := DefaultConfig()
+		cfg.RequestDurationDatastoreQueryCountBuckets = []string{"12", "-45", "66"}
+
+		err := cfg.Verify()
+		require.Error(t, err)
+	})
+}
