+++ pkg/server/server.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"net/http"
 	"slices"
 	"sort"
@@ -24,6 +23,7 @@ import (
 	httpmiddleware "github.com/openfga/openfga/pkg/middleware/http"
 	"github.com/openfga/openfga/pkg/middleware/validator"
 	"github.com/openfga/openfga/pkg/server/commands"
+	serverconfig "github.com/openfga/openfga/pkg/server/config"
 	serverErrors "github.com/openfga/openfga/pkg/server/errors"
 	"github.com/openfga/openfga/pkg/storage"
 	"github.com/openfga/openfga/pkg/storage/storagewrappers"
@@ -47,18 +47,6 @@ const (
 	AuthorizationModelIDHeader                          = "openfga-authorization-model-id"
 	authorizationModelIDKey                             = "authorization_model_id"
 	ExperimentalCheckQueryCache ExperimentalFeatureFlag = "check-query-cache"
-
-	// same values as run.DefaultConfig() (TODO break the import cycle, remove these hardcoded values and import those constants here)
-	defaultChangelogHorizonOffset           = 0
-	defaultResolveNodeLimit                 = 25
-	defaultResolveNodeBreadthLimit          = 100
-	defaultListObjectsDeadline              = 3 * time.Second
-	defaultListObjectsMaxResults            = 1000
-	defaultMaxConcurrentReadsForCheck       = math.MaxUint32
-	defaultMaxConcurrentReadsForListObjects = math.MaxUint32
-	defaultCheckQueryCacheLimit             = 10000
-	defaultCheckQueryCacheTTL               = 10 * time.Second
-	defaultCheckQueryCacheEnable            = false
 )
 
 var tracer = otel.Tracer("openfga/pkg/server")
@@ -257,18 +245,18 @@ func NewServerWithOpts(opts ...OpenFGAServiceV1Option) (*Server, error) {
 		logger:                           logger.NewNoopLogger(),
 		encoder:                          encoder.NewBase64Encoder(),
 		transport:                        gateway.NewNoopTransport(),
-		changelogHorizonOffset:           defaultChangelogHorizonOffset,
-		resolveNodeLimit:                 defaultResolveNodeLimit,
-		resolveNodeBreadthLimit:          defaultResolveNodeBreadthLimit,
-		listObjectsDeadline:              defaultListObjectsDeadline,
-		listObjectsMaxResults:            defaultListObjectsMaxResults,
-		maxConcurrentReadsForCheck:       defaultMaxConcurrentReadsForCheck,
-		maxConcurrentReadsForListObjects: defaultMaxConcurrentReadsForListObjects,
+		changelogHorizonOffset:           serverconfig.DefaultChangelogHorizonOffset,
+		resolveNodeLimit:                 serverconfig.DefaultResolveNodeLimit,
+		resolveNodeBreadthLimit:          serverconfig.DefaultResolveNodeBreadthLimit,
+		listObjectsDeadline:              serverconfig.DefaultListObjectsDeadline,
+		listObjectsMaxResults:            serverconfig.DefaultListObjectsMaxResults,
+		maxConcurrentReadsForCheck:       serverconfig.DefaultMaxConcurrentReadsForCheck,
+		maxConcurrentReadsForListObjects: serverconfig.DefaultMaxConcurrentReadsForListObjects,
 		experimentals:                    make([]ExperimentalFeatureFlag, 0, 10),
 
-		checkQueryCacheEnabled: defaultCheckQueryCacheEnable,
-		checkQueryCacheLimit:   defaultCheckQueryCacheLimit,
-		checkQueryCacheTTL:     defaultCheckQueryCacheTTL,
+		checkQueryCacheEnabled: serverconfig.DefaultCheckQueryCacheEnable,
+		checkQueryCacheLimit:   serverconfig.DefaultCheckQueryCacheLimit,
+		checkQueryCacheTTL:     serverconfig.DefaultCheckQueryCacheTTL,
 		checkCache:             nil,
 
 		requestDurationByQueryHistogramBuckets: []uint{50, 200},
